requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
google-generativeai==0.7.2
gunicorn==23.0.0
jinja2==3.1.4
itsdangerous==2.2.0